+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.0.1.1] - 2019-12-31
 ### Added
-- `param` combinator to add named parameters to selectors.
+- `param` combinator to add named parameters to selectors;
+- `CypherDSLParams` to control parameters that queries accept.
 
 ## [0.0.1.0] - 2019-12-17
 ### Changed